ajout du boutton modifier et du boutton supprimer de gestion commande

+++ APP_GRAPHIQUE_PROJET/GestionCommandes.h
@@ -33,7 +33,7 @@ namespace APPGRAPHIQUEPROJET {
 		bool check_commande_payment_means_entry();
 		bool check_commande_ID_delivery_entry();
 		bool check_commande_ID_facturation_entry();
-		bool check_commande_ID_article_entry();
+		bool check_commande_ref_entry();
 
 	protected:
 		/// <summary>
@@ -748,9 +748,8 @@ private: System::Void BOUTON_CLEAR_CLIENT_Click(System::Object^ sender, System::
 		bool isMoyenPaimentValid = check_commande_payment_means_entry();
 		bool isIDDeliveryValid = check_commande_ID_delivery_entry();
 		bool isIDFacturationValid = check_commande_ID_facturation_entry();
-		bool isIDArticleValid = check_commande_ID_article_entry();
 
-		if (isIdClientsValid & isMoyenPaimentValid & isIDDeliveryValid & isIDFacturationValid & isIDArticleValid) {
+		if (isIdClientsValid & isMoyenPaimentValid & isIDDeliveryValid & isIDFacturationValid ) {
 			this->oSvc_Commande->ajouter_Commande(Date_Livraison, 0, 0, 0, Moyen_Paiment, Client_ID, Commande_ID_Livraison, Commande_ID_Facturation);
 		}
 		
@@ -790,8 +789,15 @@ private: System::Void BOUTON_MODIFIER_COMMANDE_Click(System::Object^ sender, Sys
 	DateTime^ Date_Emission = DateTime::Parse(this->TXT_DATE_EM_COMMANDE->Text);
 
 
-	
-	this->oSvc_Commande->modifier_Commande(Commande_ID, Date_Livraison, Date_Emission, 0, 0, 0, Moyen_Paiment, Client_ID, Commande_ID_Livraison, Commande_ID_Facturation);
+	bool isIdClientsValid = check_commande_ID_cient_entry();
+	bool isMoyenPaimentValid = check_commande_payment_means_entry();
+	bool isIDDeliveryValid = check_commande_ID_delivery_entry();
+	bool isIDFacturationValid = check_commande_ID_facturation_entry();
+	bool isIDReferenceValid = check_commande_ref_entry();
+
+	if (isIdClientsValid & isMoyenPaimentValid & isIDDeliveryValid & isIDFacturationValid & isIDReferenceValid) {
+		this->oSvc_Commande->modifier_Commande(Commande_ID, Date_Livraison, Date_Emission, 0, 0, 0, Moyen_Paiment, Client_ID, Commande_ID_Livraison, Commande_ID_Facturation);
+	}
 
 	this->TXT_ID_COMMANDE->Text = "";  
 	this->TXT_ID_CLIENT_COMMANDE->Text = "";
@@ -814,7 +820,11 @@ private: System::Void BOUTON_MODIFIER_COMMANDE_Click(System::Object^ sender, Sys
 	{
 		String^ Commande_ID = this->TXT_ID_COMMANDE->Text;
 
-		this->oSvc_Commande->supprimer_Commande(Commande_ID);
+		bool isIDReferenceValid = check_commande_ref_entry();
+
+		if (isIDReferenceValid) {
+			this->oSvc_Commande->supprimer_Commande(Commande_ID);
+		}
 
 		this->TXT_ID_COMMANDE->Text = "";
